+++ packages/lib-classifier/src/plugins/tasks/experimental/TranscriptionTask/index.js
@@ -1,4 +1,4 @@
-import { default as TaskComponent } from '../DrawingTask/components/DrawingTask'
+import { default as TaskComponent } from '../../DrawingTask/components/DrawingTask'
 import { default as TaskModel } from './models/TranscriptionTask'
 import { default as AnnotationModel } from './models/TranscriptionAnnotation'
