+++ packages/lib-classifier/src/plugins/tasks/experimental/TranscriptionTask/models/TranscriptionTask.js
@@ -1,7 +1,7 @@
 import cuid from 'cuid'
 import { types } from 'mobx-state-tree'
-import Task from '../../models/Task'
-import DrawingTask from '../../DrawingTask/models/DrawingTask'
+import Task from '../../../models/Task'
+import DrawingTask from '../../../DrawingTask/models/DrawingTask'
 import { TranscriptionLineTool } from '@plugins/drawingTools/models/tools'
 import { TranscriptionLine } from '@plugins/drawingTools/models/marks'
 import TranscriptionAnnotation from './TranscriptionAnnotation'
@@ -11,7 +11,7 @@ import TranscriptionAnnotation from './TranscriptionAnnotation'
 const Transcription = types.model('Transcription', {
   activeMark: types.safeReference(TranscriptionLine),
   annotation: types.safeReference(TranscriptionAnnotation),
-  caesarKey: types.maybe(types.literal('alice')),
+  caesarKey: types.optional(types.literal('alice'), 'alice'),
   tools: types.array(TranscriptionLineTool),
   type: types.literal('transcription')
 })
